LayoutTests/fast/viewport/viewport-17-expected.txt

ALERT: viewport size 980x1078 scale 5 with limits [5, 5] and userScalable true
